[Klib] Deduplicate IrFileEntry
In Kotlin 2.1.20, the Klib format had the following features: IrFile contains one FileEntry, which is written as its own field in a protobuf message. No other FileEntry existed

In Kotlin 2.2.0, other FileEntry appeared in the new IR node IrInlinedFunctionBlock. Having lots of inlined call sites from the same source file would lead to excessive duplication of serialized FileEntry records. So, it makes sense to de-duplicate them by putting in its own table(serialized to file fileEntries.knf) and using indices to this table.

In scope of KT-73672, KLib ABI version is bumped for 2.2.0 release, so the Klib format may be amended in the same release. Backwards compatibility with KLibs having the previous ABI version must be maintained:
* Serializer should write new KLibs in a "new" way, by serializing an index to FileEntry
* Deserializer should support both new (index) and old (full FileEntry record) ways.

^KT-73826 Fixed

Merge-request: KT-MR-19515
Merged-by: Vladimir Sukharev <Vladimir.Sukharev@jetbrains.com>

+++ compiler/util-klib/src/org/jetbrains/kotlin/library/impl/KotlinLibraryLayoutImpl.kt
@@ -103,10 +103,16 @@ class FromZipIrLibraryImpl(zipped: IrLibraryLayoutImpl, zipFileSystem: FileSyste
 fun KotlinLibraryLayoutImpl.extract(file: File): File = extract(this.klib, file)
 
 private fun extract(zipFile: File, file: File) = zipFile.withZipFileSystem { zipFileSystem ->
-    val temporary = org.jetbrains.kotlin.konan.file.createTempFile(file.name)
-    zipFileSystem.file(file).copyTo(temporary)
-    temporary.deleteOnExit()
-    temporary
+    val innerFile = zipFileSystem.file(file)
+    if (innerFile.exists) {
+        val temporary = org.jetbrains.kotlin.konan.file.createTempFile(file.name)
+        innerFile.copyTo(temporary)
+        temporary.deleteOnExit()
+        temporary
+    } else {
+        // return deliberately nonexistent file name zipFile.name + ! + file.name
+        File(zipFile.path + "!" + file.path)
+    }
 }
 
 fun KotlinLibraryLayoutImpl.extractDir(directory: File): File = extractDir(this.klib, directory)
@@ -154,6 +160,8 @@ class ExtractingIrLibraryImpl(val zipped: IrLibraryLayoutImpl) :
     override val irFiles: File by lazy { zipped.extract(zipped.irFiles) }
 
     override val irDebugInfo: File by lazy { zipped.extract(zipped.irDebugInfo) }
+
+    override val irFileEntries: File by lazy { zipped.extract(zipped.irFileEntries) }
 }
 
 internal fun zippedKotlinLibraryChecks(klibFile: File) {
